Update role points if adding a role which already exists
This *probably* shouldn't be how to handle this, but it's a temporary solution to the problem stated in https://discord.com/channels/208895639164026880/689189589776203861/951929795057438740

Basically, the buttons inside the Role Gain Settings page don't actually do anything, and just return back to the Main settings page. I have no idea how to fix this, but this seems like an okay way to get around not being able to edit roles. 

Unfortunately, we can't delete roles with this approach, so we should actually fix the buttons in the Role Gain Settings page /shrug